Add an example settings file
This patch creates a copy of the existing ‘settings.php’ file intended
to be an example for the actual settings file.  One installation of
vir may use different paths for Smarty, for example.  It is slightly
annoying to see Git remind us about changes to commit whenever we need
to modify something in the settings file.  So instead we will only
provide an example settings file and tell developers that they must
create a copy of that which vir will use, and which Git will ignore.