Add find() and add() methods to the ObjectGroup trait
Using a find() method like this will reduce the amount of
redundancy the container arrays were having before and allow us
to use an enumerated array (almost like a hash!). Add() lets us
centralize this behavior if it needs to be updated in the future.

This currently adds an un-checked requirements for container items
to have a 'name', so I'll need to add verification for this in the
future.