Implement ArrayAccess for ObjectGroup class
Originally, I thought this was necessary for getting json_encode to
work, but that's not the case (that will be next via implementing
JsonSerializable). But I'm leaving this in anyway, as it effectively
solves the problem that I was previously solving with the add() function.
Subsequently, this add() function has been removed. Currently, the
offsetUnset() function doesn't do anything.